Add file logging capability
This still needs to be tested a bit. Still want to switch all of this over to the actual logging API included in the project already.
