+++ docs/com/github/tcn/plexi/gui/SwingOutputStream.html
@@ -127,7 +127,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.io.OutputStream</li>
 <li>
 <ul class="inheritance">
-<li>com.github.tcn.plexi.gui.SwingOutputStream</li>
+<li>com.github.tcn.plexi.gui.DualOutputStream</li>
 </ul>
 </li>
 </ul>
